refactor: Separate util functions
